requirements.txt
Flask==2.2.2
Werkzeug==2.3.2
pymodbus==2.5.3
pytz==2021.3
requests==2.31.0
paho-mqtt==2.1.0